+++ deno_dist/helper/ssg/index.ts
@@ -2,38 +2,7 @@ import { Buffer } from "node:buffer";
 import { inspectRoutes } from '../../helper/dev/index.ts'
 import type { Hono } from '../../hono.ts'
 import type { Env, Schema } from '../../types.ts'
-
-export const dirname = (path: string) => {
-  const splitedPath = path.split(/[\/\\]/)
-
-  return splitedPath.slice(0, -1).join('/') // Windows supports slash path
-}
-export const joinPaths = (...paths: string[]) => {
-  const nomalizedPaths: string[] = []
-  for (const path of paths) {
-    const nomalizedPath = path.replace(/(\\)/g, '/').replace(/\/$/g, '')
-    nomalizedPaths.push(nomalizedPath)
-  }
-  const resultPaths: string[] = []
-  for (let path of nomalizedPaths.join('/').split('/')) {
-    // Handle `..` or `../`
-    if (path === '..') {
-      if (resultPaths.length === 0) {
-        resultPaths.push('..')
-      } else {
-        resultPaths.pop()
-      }
-      continue
-    } else {
-      // Handle `.` or `./`
-      path = path.replace(/^\./g, '')
-    }
-    if (path !== ''){
-      resultPaths.push(path)
-    }
-  }
-  return resultPaths.join('/')
-}
+import { joinPaths, dirname } from './utils.ts'
 
 export interface FileSystemModule {
   writeFile(path: string, data: string | Buffer): Promise<void>
